requirements.txt
pydantic>=2.6
python-dotenv>=1.0
pyyaml>=6.0
httpx>=0.27           # Semantic Scholar client uses httpx directly
tavily-python>=0.3.7  # Stage 3/4: supplier search + price scraping
openai>=1.40
anthropic>=0.34

# Tests
pytest>=8.0
